Go back to having Gamepad Mode in the same class
